CXFLW-1642 Added code to cehck for deleting project and CXFLW-1641 SCA result

+++ src/main/java/com/checkmarx/sdk/utils/scanner/client/ScaClientHelper.java
@@ -1698,7 +1698,11 @@ public class ScaClientHelper extends ScanClientHelper implements IScanClientHelp
             log.info("Getting project by name for deleting project: '{}'", projectName);
             String resolvedProjectId = getRiskManagementProjectId(projectName);
             log.info("deleting CxSCA project according to project id {}",resolvedProjectId);
-            httpClient.deleteRequest(String.format(PROJECTS_BY_ID, resolvedProjectId), HttpStatus.SC_NO_CONTENT, "delete a project");
+            if(resolvedProjectId!=null || !resolvedProjectId.isEmpty()){
+                httpClient.deleteRequest(String.format(PROJECTS_BY_ID, resolvedProjectId), HttpStatus.SC_NO_CONTENT, "delete a project");
+            }else{
+                log.info("deleting CxSCA project {} does not exist.",projectName);
+            }
         }catch (IOException e){
             log.error("Error while deleting project");
             throw new RuntimeException(e);
